requirements.txt
discord.py
asyncpg
aiohttp
noise
pillow
imageio
numpy
scipy
jellyfish
regex
matplotlib
git+https://github.com/AndreyATGB/Keras-RNN.git
opencv-python-headless # Will NOT work, H264 support is expected
tensorflow < 1.15 # 1.15 does not work (it does?), need 1.14 or lower
tensorflow_hub
keras < 2.4 # 2.4 requires TF 2.2
